requirements.txt
# Core training (bin/train.sh, bin/train.py)
torch
hyperpyyaml
pyyaml
tqdm
tensorboard
webdataset
